ready to submit

+++ src/app/modules/admin/admin.service.ts
@@ -1,17 +1,28 @@
+import customError from "../../errors/customError";
 import { Blog } from "../blog/blog.model";
 import { User } from "../user/user.model";
 
 const blockUserFromDB = async (_id: string) => {
+  const isUserExist = await Blog.findById(_id);
+  if (!isUserExist) {
+    throw new customError(404, "User not found");
+  }
+
   const result = await User.findOneAndUpdate({ _id }, { isBlocked: true });
   return result;
 };
 
 const deleteBlogFromDB = async (_id: string) => {
+  const isUserExist = await Blog.findById(_id);
+  if (!isUserExist) {
+    throw new customError(404, "User not found");
+  }
+
   const result = await Blog.deleteOne({ _id });
   return result;
 };
 
 export const adminServices = {
   blockUserFromDB,
-  deleteBlogFromDB
+  deleteBlogFromDB,
 };
